MusicStore/src/app/genres/genre-detail/genre-detail.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { GenreDetailComponent } from './genre-detail.component';

describe('Component: GenreDetail', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [GenreDetailComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([GenreDetailComponent],
      (component: GenreDetailComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(GenreDetailComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(GenreDetailComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <ms-genre-detail></ms-genre-detail>
  `,
  directives: [GenreDetailComponent]
})
class GenreDetailComponentTestController {
}

